# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: ContentFlow API
  description: AI SEO content generation platform.
  version: 0.1.0
servers:
  - url: /
    description: Same origin as the docs (works on any host/port)
  - url: https://multiagent-seo.8cell.tech
    description: Staging
...
paths:
...
  /clients/{id}/notes:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: addClientNote
      security:
        - bearerAuth: [ admin ]
      tags: [ Clients ]
      summary: Add a manual call/contact log entry to a client's card — free text staff types in after a call, there's no phone system integration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddClientNoteRequest"
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientNote"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /vault-entries:
    get:
      operationId: listVaultEntries
      security:
        - bearerAuth: [ admin ]
      tags: [ Vault ]
      summary: List saved password/credential entries
      responses:
        "200":
          description: All entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultEntry"
    post:
      operationId: createVaultEntry
      security:
        - bearerAuth: [ admin ]
      tags: [ Vault ]
      summary: Add a password/credential entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVaultEntryRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
  /vault-entries/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: updateVaultEntry
      security:
        - bearerAuth: [ admin ]
      tags: [ Vault ]
      summary: Update a password/credential entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVaultEntryRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteVaultEntry
      security:
        - bearerAuth: [ admin ]
      tags: [ Vault ]
      summary: Delete a password/credential entry
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      description: RFC 7807 problem details
      required: [ title, status ]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
    ClientNote:
      type: object
      required: [ id, text, created_by, created_at ]
      properties:
        id:
          type: string
        text:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
...
    AddClientNoteRequest:
      type: object
      required: [ text ]
      properties:
        text:
          type: string
    VaultEntry:
      type: object
      required: [ id, title, url, username, password, notes, createdBy, createdAt, updatedAt ]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
        username:
          type: string
        password:
          type: string
        notes:
          type: string
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateVaultEntryRequest:
      type: object
      required: [ title ]
      properties:
        title:
          type: string
          x-oapi-codegen-extra-tags:
            validate: "required,min=1,max=200"
        url:
          type: string
        username:
          type: string
        password:
          type: string
        notes:
          type: string
    UpdateVaultEntryRequest:
      type: object
      properties:
        title:
          type: string
          x-oapi-codegen-extra-tags:
            validate: "omitempty,min=1,max=200"
        url:
          type: string
        username:
          type: string
        password:
          type: string
        notes:
          type: string
...